Add files via upload
so i set up a basic setup for our gui, i think button1 is what does a select * from admin; query and displays the results to the table view box.
pretty much, pick something to work off of such as organizing the gui a little cleaner or making the modules.
you could also figure out the login and logout feature. check those videos i sent in the texts. just work on it, if we get something decent by tuesday we should be a little better off.
godspeed and good luck.